bruh2

+++ task3/Spy.hpp
@@ -8,26 +8,26 @@ template <class T> class Spy {
 private:
   T value_;
   void *logger_ = nullptr;
-  size_t counter = 0;
-  size_t sz = 0;                           //
-  void (*logpt)(void *, size_t) = nullptr; //
+  unsigned int counter = 0;
+  unsigned int sz = 0;                           //
+  void (*logpt)(void *, unsigned int) = nullptr; //
   void (*destrpt)(void *) = nullptr;
   void *(*coptr)(void *) = nullptr;
-  using BigLogType = void (*)(void *, size_t);
+  using BigLogType = void (*)(void *, unsigned int);
 
 public:
   class LogInvoke {
   private:
     Spy<T> *spy_ptr;
-    size_t i = 0;
+    unsigned int i = 0;
     T *t;
     void *logger;
     BigLogType log;
-    size_t *curr;
+    unsigned int *curr;
 
   public:
     // T *operator->() { return &spy_ptr->value_; }
-    LogInvoke(T *t, void *logger, BigLogType log, size_t *counter)
+    LogInvoke(T *t, void *logger, BigLogType log, unsigned int *counter)
         : t(t), logger(logger), log(log), curr(counter) {}
 
     ~LogInvoke() {
@@ -151,14 +151,14 @@ public:
     return *this;
   }
 
-  template <std::invocable<size_t> Log_>
+  template <std::invocable<unsigned int> Log_>
     requires(std::destructible<T>) && // unnecessary?
             (std::destructible<std::remove_cvref_t<Log_>>)
   void setHelperFunc(Log_ &&logger) {
     // auto nextres = std::forward<Log_>(logger);
     logger_ = new std::remove_cvref_t<Log_>(std::forward<Log_>(logger));
 
-    logpt = +[](void *self, size_t qnt) {
+    logpt = +[](void *self, unsigned int qnt) {
       std::invoke(*static_cast<std::remove_cvref_t<Log_> *>(self), qnt);
     };
 
@@ -167,7 +167,7 @@ public:
     };
   }
 
-  template <std::invocable<size_t> Log_> // invoc
+  template <std::invocable<unsigned int> Log_> // invoc
     requires(!std::copyable<T>) && (std::destructible<T>) &&
             (std::destructible<std::remove_cvref_t<Log_>>) &&
             (std::movable<T>) &&
@@ -175,7 +175,7 @@ public:
   void setLogger(Log_ &&logger) {
     logger_ = new std::remove_cvref_t<Log_>(std::forward<Log_>(logger));
 
-    logpt = +[](void *self, size_t qnt) {
+    logpt = +[](void *self, unsigned int qnt) {
       std::invoke(*static_cast<std::remove_cvref_t<Log_> *>(self), qnt);
     };
 
@@ -184,14 +184,14 @@ public:
     };
   }
 
-  template <std::invocable<size_t> Log_>
+  template <std::invocable<unsigned int> Log_>
     requires(std::copyable<T>) && (std::destructible<T>) &&
             (std::destructible<std::remove_cvref_t<Log_>>) &&
             (std::copyable<std::remove_cvref_t<Log_>>)
   void setLogger(Log_ &&logger) {
     logger_ = new std::remove_cvref_t<Log_>(std::forward<Log_>(logger));
 
-    logpt = +[](void *self, size_t qnt) {
+    logpt = +[](void *self, unsigned int qnt) {
       std::invoke(*static_cast<std::remove_cvref_t<Log_> *>(self), qnt);
     };
